Move contains trait to contains.rs
